\name{03.Morris}
\alias{03.Morris}
\alias{Morris}
\alias{morris}
\title{Converted Matlab functions from Morris and Doak(2002)} 
\description{ 
Chapter 3
\describe{
\item{ \code{\link{countCDFxt}} }{ Box 3.4. Calculate count-based
    extinction probabilities and bootstrap confidence intervals.  Includes some functions
  confidence intervals  }
}

Chapter 7
\describe{
\item{ \code{\link{stoch.projection}} }{  Box 7.3.  Project stochastic growth from a sequence of matrices }
\item{ \code{\link{stoch.growth.rate}} }{ Box 7.4.  Calculate the log stochastic growth rate by Tuljapukar's approximation and by simulation   }
\item{ \code{\link{stoch.quasi.ext}} }{  Box 7.5.  Estimate quasi-extinction threshold    }
}

Chapter 8
\describe{
\item{ \code{\link{betaval}} }{  Box 8.3. Generate beta-distributed random numbers  }
\item{ \code{\link{lnorms}} }{ Box 8.4.  Generate random lognormal values   }
\item{ \code{\link{stretchbetaval}} }{ Box 8.5.  Generate stretched
    beta-distributed random numbers  }
\item{ \code{\link{vitalsim}} }{  Box 8.10.  Calculate stochastic growth rate and extinction time CDF using
  vital rates  }
\item{ \code{\link{multiresultm}} }{ Box 8.11.  Incorporate demographic stochasticity into population
projections }
}


}


\keyword{documentation}
