\name{Morris}
\alias{03.Morris}
\alias{Morris}
\title{Converted Matlab functions from Morris and Doak (2002)}
\description{
  \bold{Chapter 3} \describe{ \item{ }{ Table 3.1. Grizzly
  bear population counts.  The example includes code to
  calculate mean, variance and confidence intervals using
  regression and other procedures
  }\item{list(list("grizzly"))}{ Table 3.1. Grizzly bear
  population counts. The example includes code to calculate
  mean, variance and confidence intervals using regression
  and other procedures }\item{ }{ Table 3.1. Grizzly bear
  population counts.  The example includes code to
  calculate mean, variance and confidence intervals using
  regression and other procedures } \item{ }{ Box 3.3.
  Count-based extinction time cumulative distribution
  function }\item{list(list("extCDF"))}{ Box 3.3.
  Count-based extinction time cumulative distribution
  function }\item{ }{ Box 3.3. Count-based extinction time
  cumulative distribution function } \item{ }{ Box 3.4.
  Count-based extinction probabilities with bootstrap
  confidence intervals }\item{list(list("countCDFxt"))}{
  Box 3.4. Count-based extinction probabilities with
  bootstrap confidence intervals }\item{ }{ Box 3.4.
  Count-based extinction probabilities with bootstrap
  confidence intervals } }
}
\details{
  \bold{Chapter 7} \describe{ \item{ }{ Box 7.3.  Project
  stochastic growth from a sequence of matrices
  }\item{list(list("stoch.projection"))}{ Box 7.3. Project
  stochastic growth from a sequence of matrices }\item{ }{
  Box 7.3. Project stochastic growth from a sequence of
  matrices } \item{ }{ Box 7.4. Calculate the log
  stochastic growth rate by Tuljapukar's approximation and
  by simulation }\item{list(list("stoch.growth.rate"))}{
  Box 7.4.  Calculate the log stochastic growth rate by
  Tuljapukar's approximation and by simulation }\item{ }{
  Box 7.4.  Calculate the log stochastic growth rate by
  Tuljapukar's approximation and by simulation } \item{ }{
  Box 7.5.  Estimate quasi-extinction threshold
  }\item{list(list("stoch.quasi.ext"))}{ Box 7.5. Estimate
  quasi-extinction threshold }\item{ }{ Box 7.5.  Estimate
  quasi-extinction threshold } }

  \bold{Chapter 8} \describe{ \item{ }{ Box 8.2. Kendall's
  method to correct for sampling variation
  }\item{list(list("Kendall"))}{ Box 8.2. Kendall's method
  to correct for sampling variation }\item{ }{ Box 8.2.
  Kendall's method to correct for sampling variation }
  \item{ }{ Box 8.3. Generate beta-distributed random
  numbers }\item{list(list("betaval"))}{ Box 8.3. Generate
  beta-distributed random numbers }\item{ }{ Box 8.3.
  Generate beta-distributed random numbers } \item{ }{ Box
  8.4.  Generate random lognormal values
  }\item{list(list("lnorms"))}{ Box 8.4.  Generate random
  lognormal values }\item{ }{ Box 8.4.  Generate random
  lognormal values } \item{ }{ Box 8.5.  Generate stretched
  beta-distributed random numbers
  }\item{list(list("stretchbetaval"))}{ Box 8.5.  Generate
  stretched beta-distributed random numbers }\item{ }{ Box
  8.5.  Generate stretched beta-distributed random numbers
  } \item{ }{ Box 8.10.  Calculate stochastic growth rate
  and extinction time CDF using vital rates
  }\item{list(list("vitalsim"))}{ Box 8.10.  Calculate
  stochastic growth rate and extinction time CDF using
  vital rates }\item{ }{ Box 8.10.  Calculate stochastic
  growth rate and extinction time CDF using vital rates }
  \item{ }{ Box 8.11.  Incorporate demographic
  stochasticity into population projections
  }\item{list(list("multiresultm"))}{ Box 8.11.
  Incorporate demographic stochasticity into population
  projections }\item{ }{ Box 8.11.  Incorporate demographic
  stochasticity into population projections } }

  \bold{Chapter 9} \describe{ \item{ }{ Box 9.1. Vital rate
  sensitivity and
  elasticity}\item{list(list("vitalsens"))}{ Box 9.1. Vital
  rate sensitivity and elasticity}\item{ }{ Box 9.1. Vital
  rate sensitivity and elasticity} }
}
\keyword{documentation}

