% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var2.R
\name{var2}
\alias{var2}
\title{Variance matrix}
\usage{
var2(x)
}
\arguments{
\item{x}{A list of two or more matrices}
}
\value{
A matrix containing variances
}
\description{
Calculates the variances from a list of matrices
}
\examples{
var2(hudsonia)

}
\author{
Chris Stubben
}
