\name{02.Caswell}
\alias{02.Caswell}
\alias{Caswell}
\alias{caswell}
\title{Converted Matlab functions from Caswell (2001) } 
\description{ 
Chapter 5.  Events in the Life Cycle
\describe{
\item{ \code{\link{fundamental.matrix}} }{   section 5.3.1.   Calculate age-specific survival from a
stage classified matrix using the fundamental matrix N }
\item{ \code{\link{net.reproductive.rate}} }{   section 5.3.4.  Calculate the net reproductive rate of a stage classified matrix using the dominant eigenvalue of the matrix R.  }
\item{ \code{\link{generation.time}} }{ section 5.3.5.   Calculate the generation time of a stage-classified matrix  }
}

Additional functions
\describe{
\item{ \code{\link{eigen.analysis}} }{ section 4.8, 9.1, and 9.2.
    Eigenvalues and eigenvectors to calculate population growth rate,
    sensitivities and elasticities  }
\item{ \code{\link{QPmat}} }{ section 6.2.2.  Construct a projection
    matrix from a time series of  individuals per stage using Wood's quadratic programming method  }
\item{ \code{\link{multiresultm}} }{ section 15.1.3.  Incorporate demographic stochasticity into population
projections }

}







}
\author{Chris Stubben}

\keyword{documentation}
