% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head2.R
\name{head2}
\alias{head2}
\title{Return the first and last part of a matrix or dataframe}
\usage{
head2(x, head = 3, tail = 1, dotrows = 1)
}
\arguments{
\item{x}{A matrix or dataframe}

\item{head}{The number of first rows}

\item{tail}{The number of last rows}

\item{dotrows}{The number of rows of dots}
}
\value{
A smaller object with first and last rows only
}
\description{
Returns the first and last rows using output from both \code{\link{head}} and
\code{\link{tail}} and separates the two parts with dots. Useful for viewing
ordered datasets such as longitudinal census data.
}
\examples{
head2(aq.trans)

}
\author{
Chris Stubben
}
