% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudcorrs}
\alias{hudcorrs}
\title{Correlation matrices for Hudsonia vital rates}
\format{
A list with 2 correlation matrices, corrin (within year correlation)
and corrout (between year correlation).
}
\source{
The correlation matrices in Morris and Doak 2002 include some
correlations > 1.  A corrected set of correlations was sent by the D. Doak on
8/4/2007.
}
\usage{
hudcorrs
}
\description{
Within year and between year correlation matrices from \emph{Hudsonia
montana} vital rates.  Correlations were calculated from first 13 growth and
survival rates only, since fertility rates vary little.
}
\examples{
hudcorrs
}
\references{
Morris, W. F., and D. F. Doak. 2002. Quantitative conservation
biology: Theory and practice of population viability analysis. Sinauer,
Sunderland, Massachusetts, USA.
}
\seealso{
\code{\link{vitalsim}}
}
\keyword{datasets}
