% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudvrs}
\alias{hudvrs}
\title{Best Kendall estimates of Hudsonia vital rate means and variances}
\format{
A data frame with 24 rows and 2 columns:
\describe{
  \item{\code{mean}}{ vital rate means}
  \item{\code{var}}{ vital rate variances}
}
}
\source{
Data  listed in Box 8.10 for the \code{\link{vitalsim}} function.
See also Table 8.5 in Morris and Doak (2002).
}
\usage{
hudvrs
}
\description{
Best Kendall estimates of vital rate means (9 growth, 4 survival, and 11
fertility rates) for \emph{Hudsonia montana}.
}
\examples{
hudvrs
hudmxdef(hudvrs$mean)
}
\references{
Morris, W. F., and D. F. Doak. 2002. Quantitative conservation
biology: Theory and practice of population viability analysis. Sinauer,
Sunderland, Massachusetts, USA.
}
\keyword{datasets}
