% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teasel}
\alias{teasel}
\title{Projection matrix for teasel}
\format{A 6 x 6 matrix}
\source{
Example 5.2 in Caswell 2001.
}
\usage{
teasel
}
\description{
Projection matrix with six stage classes for the plant teasel
}
\examples{
data(teasel)
image2(teasel, mar = c(1, 3.5, 5, 1), box.offset = .1)
title("Teasel projection matrix", line = 2.5)
# fertilities for a monocarpic plant in a prebreeding census in last column
splitA(teasel, r = 1:6, c = 6)
lambda(teasel)
}
\references{
Caswell, H. 2001. Matrix population models. Construction,
Analysis and interpretation. 2nd ed. Sinauer, Sunderland, Massachusetts.
}
\keyword{datasets}
