% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whale}
\alias{whale}
\title{Projection matrix for killer whale}
\format{
A 4 x 4 matrix
}
\source{
Example 5.1 in Caswell (2001)
}
\usage{
whale
}
\description{
Projection matrix for killer whales with 4 size classes: yearling, juvenile,
mature and post-reproductive
}
\examples{
whale
splitA(whale)
lambda(whale)
sensitivity(whale)
# plot sensitivity
matplot2(sensitivity(whale),
  type = "b", legend = "topleft", ltitle = "Fate",
  main = "Killer Whale sensitivity"
)
}
\references{
Caswell, H. 2001. Matrix population models. Construction,
Analysis and interpretation. 2nd ed. Sinauer, Sunderland, Massachusetts.
}
\keyword{datasets}
